Revert "Feature/login screen (#159)"
This reverts commit 6a3b41d98d7ceefda3a00dfc458e67e6871656e0.
