--- lib/assets/components/corner_clip_path.dart
@@ -1,21 +1,0 @@
-import 'package:flutter/material.dart';
-
-class BackgroundClipper extends CustomClipper<Path> {
-  @override
-  Path getClip(Size size) {
-    var path = Path();
-    path.moveTo(size.width * 0.2, 0);
-    path.lineTo(size.width, 0);
-    path.lineTo(size.width, size.height);
-    path.quadraticBezierTo(size.width * 0.75, size.height * 0.9,
-        size.width * 0.65, size.height * 0.6);
-    path.quadraticBezierTo(
-        size.width * 0.5, size.height * 0.1, size.width * 0.2, 0);
-    return path;
-  }
-
-  @override
-  bool shouldReclip(CustomClipper<Path> oldClipper) {
-    return true;
-  }
-}
