+++ lib/assets/components/inputs.dart
@@ -16,8 +16,6 @@ class CustomTextFormField extends StatelessWidget {
   final CustomFormFieldStyle style;
   final String hintText;
   final TextCapitalization textCapitalization;
-  final Color backgroundColor;
-  final Color hintTextColor;
 
   CustomTextFormField({
     @required this.autofocus,
@@ -26,8 +24,6 @@ class CustomTextFormField extends StatelessWidget {
     @required this.onSaved,
     @required this.style,
     @required this.hintText,
-    this.hintTextColor,
-    this.backgroundColor,
     this.textCapitalization,
   });
 
@@ -40,20 +36,21 @@ class CustomTextFormField extends StatelessWidget {
       autofocus: autofocus,
       validator: validator,
       onSaved: onSaved,
-      style: textStyleFrom(Theme.of(context).primaryTextTheme.headline5,
-          color: Colors.grey),
+      style: textStyleFrom(
+        Theme.of(context).primaryTextTheme.headline5,
+        color: Colors.white,
+      ),
       textCapitalization: textCapitalization ?? TextCapitalization.none,
       decoration: InputDecoration(
         contentPadding: EdgeInsets.symmetric(vertical: 15.0, horizontal: 15),
         filled: true,
-        fillColor: this.backgroundColor ?? Color.fromRGBO(221, 221, 221, 0.2),
+        fillColor: Color.fromRGBO(221, 221, 221, 0.2),
         hintText: hintText,
         hintStyle: TextStyle(
-          color: hintTextColor ??
-              colorFrom(
-                Theme.of(context).primaryColor,
-                opacity: 0.5,
-              ),
+          color: colorFrom(
+            Theme.of(context).primaryColor,
+            opacity: 0.5,
+          ),
         ),
       ),
     ));
