+++ lib/models/Notification.dart
@@ -50,9 +50,7 @@ class InternalNotification {
   Widget getBodyWidget(context) {
     return Markdown(
       data: body,
-      onTapLink: (String text, String href, String title) {
-        _navigationService.launchLink(href);
-      },
+      onTapLink: (String text, String href, String title) {_navigationService.launchLink(href);},
       styleSheet: MarkdownStyleSheet.fromTheme(Theme.of(context)).copyWith(
         p: textStyleFrom(
           Theme.of(context).primaryTextTheme.bodyText1,
